+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/forkchoice/ForkChoiceUpdatedResultSubscriber.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2022 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.statetransition.forkchoice;
+
+import java.util.Optional;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.spec.executionengine.ForkChoiceState;
+import tech.pegasys.teku.spec.executionengine.ForkChoiceUpdatedResult;
+
+public interface ForkChoiceUpdatedResultSubscriber {
+  void onForkChoiceUpdatedResult(
+      final ForkChoiceUpdatedResultNotification forkChoiceUpdatedResultNotification);
+
+  class ForkChoiceUpdatedResultNotification {
+    private final ForkChoiceState forkChoiceState;
+    private final SafeFuture<Optional<ForkChoiceUpdatedResult>> forkChoiceUpdatedResult;
+
+    public ForkChoiceUpdatedResultNotification(
+        ForkChoiceState forkChoiceState,
+        SafeFuture<Optional<ForkChoiceUpdatedResult>> forkChoiceUpdatedResult) {
+      this.forkChoiceState = forkChoiceState;
+      this.forkChoiceUpdatedResult = forkChoiceUpdatedResult;
+    }
+
+    public ForkChoiceState getForkChoiceState() {
+      return forkChoiceState;
+    }
+
+    public SafeFuture<Optional<ForkChoiceUpdatedResult>> getForkChoiceUpdatedResult() {
+      return forkChoiceUpdatedResult;
+    }
+  }
+}
